Modified ACES Tonemapping improved shadow mapping improved overlay

+++ Bloom.cpp
@@ -12,7 +12,7 @@ void Bloom::_init(int width, int height) {
 		glBindFramebuffer(GL_FRAMEBUFFER, this->FBO[i]);
 
 		glBindTexture(GL_TEXTURE_2D, this->colorBuffers[i]);
-		glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
+		glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
 		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
 		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 		glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
